Refactor shared math utilities into modules
